GET /event/{eventId} has been moved to the EventsApi and in not used anymore to search an event in a room
